+++ cmd/main.go
@@ -196,6 +196,14 @@ func main() {
 		os.Exit(1)
 	}
 
+	if err := (&controller.ImageBuildReconciler{
+		Client:  mgr.GetClient(),
+		Clients: clientPool,
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "ImageBuild")
+		os.Exit(1)
+	}
+
 	if err := (&controller.CobblerSystemReconciler{
 		Client:  mgr.GetClient(),
 		Cobbler: clientPool,